+++ src/js/gutenberg-overrides/packages/block-library/CHANGELOG.md
@@ -1,10 +1,10 @@
 # block-library changelog
 
-## 2.7.0 2018-10-25
+## 3.0.0
 
 ### Changed
 
-- 'cover-image' to 'cover' ([packages/block-library/build-module/cover](https://github.com/front/gutenberg-js/blob/v2.7.0/src/js/gutenberg-overrides/packages/block-library/build-module/cover))
+- 'cover-image' to 'cover' ([packages/block-library/build-module/cover](https://github.com/front/gutenberg-js/blob/v3.0.0/src/js/gutenberg-overrides/packages/block-library/build-module/cover))
 
 ## 2.0.0 2018-08