+++ src/js/gutenberg-overrides/packages/edit-post/CHANGELOG.md
@@ -1,6 +1,6 @@
 # edit-post changelog
 
-## 2.7.0 2018-10-25
+## 3.0.0
 
 ### Added